Some reorganization, more code, added codeblocks to docker
