+++ docker/dockerfiles/main.docker
@@ -77,6 +77,7 @@ RUN mv /bin/sh /bin/sh-old && \
         libx11-dev \
         tzdata \
         freeglut3-dev \
+        codeblocks \
         lsb-release && \
     apt-get clean && \
     rm -rf /var/lib/apt/lists/*
